+++ src/controllers/index.ts
@@ -15,4 +15,3 @@ export * from "./BlogController";
 export * from "./PaymentLemonSqueezyController";
 export * from "./contactController";
 export * from "./FaqController";
-export * from "./SearchMembersController"
