+++ src/routes/index.ts
@@ -17,4 +17,3 @@ export * from "./testimonial";
 export * from "./updateOrg";
 export * from "./user";
 export * from "./faq";
-export * from "./searchMemberRoutes"
