--- src/routes/searchMemberRoutes.ts
@@ -1,8 +1,0 @@
-import { Router } from "express";
-import { searchOrganizationMembers } from "../controllers/SearchMembersController"
-
-const searchOrganizationMembersRouter = Router();
-
-searchOrganizationMembersRouter.post("/members/search", searchOrganizationMembers);
-
-export { searchOrganizationMembersRouter };
