+++ src/services/index.ts
@@ -10,4 +10,3 @@ export * from "./organization.services";
 export * from "./payment/flutter.service";
 export * from "./contactService";
 export * from "./faq.services";
-export * from "./SearchMembersServices";
